Refactor: Split `TeamsRepository` interface into finance and membership seams
Extracted finance and membership methods from `TeamsRepository` into new interfaces `TeamsFinancesRepository` and `TeamsMembershipRepository`.
Updated `TeamsRepository` to extend these new interfaces to keep implementation unchanged.
Updated `EnterprisesViewModel`, `EnterprisesFinancesViewModel`, and `RequestsViewModel` to use the narrower interfaces.
Updated DI to inject these narrow interfaces as Singletons.
