+++ app/src/main/java/org/ole/planet/myplanet/di/RepositoryModule.kt
@@ -46,6 +46,8 @@ import org.ole.planet.myplanet.repository.SyncRepositoryImpl
 import org.ole.planet.myplanet.repository.TagsRepository
 import org.ole.planet.myplanet.repository.TagsRepositoryImpl
 import org.ole.planet.myplanet.repository.TeamsRepository
+import org.ole.planet.myplanet.repository.TeamsFinancesRepository
+import org.ole.planet.myplanet.repository.TeamsMembershipRepository
 import org.ole.planet.myplanet.repository.TeamsRepositoryImpl
 import org.ole.planet.myplanet.repository.TeamsSyncRepository
 import org.ole.planet.myplanet.repository.UploadRepository
@@ -163,4 +165,16 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun bindSyncRepository(impl: SyncRepositoryImpl): SyncRepository
+}
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class TeamsSubRepositoryModule {
+    @Binds
+    @Singleton
+    abstract fun bindTeamsFinancesRepository(impl: TeamsRepositoryImpl): TeamsFinancesRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindTeamsMembershipRepository(impl: TeamsRepositoryImpl): TeamsMembershipRepository
 }
