+++ app/src/main/java/org/ole/planet/myplanet/repository/TeamsRepository.kt
@@ -69,7 +69,7 @@ data class TeamUploadData(
     val imageName: String? = null
 )
 
-interface TeamsRepository {
+interface TeamsRepository : TeamsFinancesRepository, TeamsMembershipRepository {
     suspend fun getAllActiveTeams(): List<MyTeam>
     suspend fun getMyTeamsFlow(userId: String): Flow<List<MyTeam>>
     suspend fun getTeamSummaries(userId: String?): List<TeamSummary>
@@ -94,12 +94,6 @@ interface TeamsRepository {
 
     suspend fun getTeamNamesByIds(ids: List<String>): Map<String, String>
     fun getTasksFlow(userId: String?): Flow<List<TeamTask>>
-    suspend fun isMember(userId: String?, teamId: String): Boolean
-    suspend fun isTeamLeader(teamId: String, userId: String?): Boolean
-    suspend fun hasPendingRequest(teamId: String, userId: String?): Boolean
-    suspend fun requestToJoin(teamId: String, userId: String?, userPlanetCode: String?, teamType: String?)
-    suspend fun leaveTeam(teamId: String, userId: String?)
-    suspend fun removeMember(teamId: String, userId: String)
     suspend fun addResourceLinks(teamId: String, resources: List<TeamResourceDto>, userId: String?)
     suspend fun removeResourceLink(teamId: String, resourceId: String)
     suspend fun deleteTask(taskId: String)
@@ -110,11 +104,6 @@ interface TeamsRepository {
     suspend fun getPendingTasksForUser(userId: String, start: Long, end: Long): List<TeamTask>
     suspend fun markTasksNotified(taskIds: Collection<String>)
     suspend fun getTasksByTeamId(teamId: String): Flow<List<TeamTask>>
-    suspend fun getReportsFlow(teamId: String): Flow<List<MyTeam>>
-    suspend fun exportReportsAsCsv(reports: List<MyTeam>, teamName: String): String
-    suspend fun addReport(report: FinanceReportParams)
-    suspend fun updateReport(reportId: String, payload: FinanceReportParams)
-    suspend fun archiveReport(reportId: String)
     suspend fun logTeamVisit(teamId: String, userName: String?, userPlanetCode: String?,
         userParentCode: String?, teamType: String?
     )
@@ -126,26 +115,9 @@ interface TeamsRepository {
         teamId: String, name: String, description: String, services: String, rules: String,
         teamType: String, isPublic: Boolean, createdBy: String
     ): Boolean
-    suspend fun getTeamTransactionsWithBalance(
-        teamId: String, startDate: Long? = null,
-        endDate: Long? = null, sortAscending: Boolean = false
-    ): Flow<List<Transaction>>
-    suspend fun createTransaction(
-        teamId: String, type: String, note: String, amount: Int, date: Long,
-        parentCode: String?, planetCode: String?,
-        imageName: String? = null, imageData: ByteArray? = null
-    ): Result<Unit>
-    suspend fun respondToMemberRequest(teamId: String, userId: String, accept: Boolean): Result<Unit>
     suspend fun getTeamType(teamId: String): String?
-    suspend fun getJoinedMembers(teamId: String): List<UserEntity>
-    suspend fun refreshJoinedMembersForLogin(teamId: String): List<UserEntity>
-    suspend fun getJoinedMembersWithVisitInfo(teamId: String): List<JoinedMemberData>
-    suspend fun getJoinedMemberCount(teamId: String): Int
     suspend fun getAssignee(userId: String): UserEntity?
-    suspend fun getRequestedMembers(teamId: String): List<UserEntity>
     suspend fun isTeamNameExists(name: String, type: String, excludeTeamId: String? = null): Boolean
-    suspend fun updateTeamLeader(teamId: String, newLeaderId: String): Boolean
-    suspend fun getNextLeaderCandidate(teamId: String, excludeUserId: String?): UserEntity?
     suspend fun getTeamCreator(teamId: String): String?
     suspend fun getAvailableResourcesToAdd(teamId: String): List<MyLibrary>
